+++ subprojects/launcher/src/test/groovy/org/gradle/launcher/cli/ReleaseFeaturesTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package org.gradle.docs.releasefeatures
+package org.gradle.launcher.cli
 
-import org.gradle.docs.SystemPropertyFiles
 import spock.lang.Specification
 
 class ReleaseFeaturesTest extends Specification {
 
     def "release features must follow conventions"() {
         given:
-        def featuresFile = SystemPropertyFiles.get("org.gradle.docs.releasefeatures")
+        def featuresFile = ReleaseFeaturesTest.getClassLoader().getResourceAsStream("release-features.txt")
 
         when:
         def featuresText = featuresFile.text
